move triehasher
