+++ crates/primitives/Cargo.toml
@@ -54,10 +54,6 @@ thiserror.workspace = true
 zstd = { version = "0.13", features = ["experimental"], optional = true }
 roaring = "0.10.2"
 
-# `test-utils` feature
-hash-db = { version = "~0.15", optional = true }
-plain_hasher = { version = "0.2", optional = true }
-
 # arbitrary utils
 arbitrary = { workspace = true, features = ["derive"], optional = true }
 proptest = { workspace = true, optional = true }
@@ -79,9 +75,6 @@ serde_json.workspace = true
 test-fuzz.workspace = true
 toml.workspace = true
 triehash = "0.8"
-
-hash-db = "~0.15"
-plain_hasher = "0.2"
 
 sucds = "0.8.1"
 
@@ -124,7 +117,7 @@ optimism = [
     "revm/optimism",
 ]
 alloy-compat = ["alloy-rpc-types"]
-test-utils = ["dep:plain_hasher", "dep:hash-db"]
+test-utils = []
 
 [[bench]]
 name = "recover_ecdsa_crit"
@@ -133,11 +126,6 @@ harness = false
 [[bench]]
 name = "validate_blob_tx"
 required-features = ["arbitrary", "c-kzg"]
-harness = false
-
-[[bench]]
-name = "trie_root"
-required-features = ["arbitrary", "test-utils"]
 harness = false
 
 [[bench]]
