+++ crates/trie/common/Cargo.toml
@@ -28,6 +28,10 @@ serde.workspace = true
 itertools.workspace = true
 nybbles = { workspace = true, features = ["serde", "rlp"] }
 
+# `test-utils` feature
+hash-db = { version = "~0.15", optional = true }
+plain_hasher = { version = "0.2", optional = true }
+
 [dev-dependencies]
 arbitrary = { workspace = true, features = ["derive"] }
 assert_matches.workspace = true
@@ -36,3 +40,8 @@ proptest-derive.workspace = true
 serde_json.workspace = true
 test-fuzz.workspace = true
 toml.workspace = true
+hash-db = "~0.15"
+plain_hasher = "0.2"
+
+[features]
+test-utils = ["dep:plain_hasher", "dep:hash-db"]
