+++ crates/trie/common/src/lib.rs
@@ -32,6 +32,8 @@ mod subnode;
 pub use subnode::StoredSubNode;
 
 mod proofs;
+#[cfg(any(test, feature = "test-utils"))]
+pub use proofs::triehash;
 pub use proofs::{AccountProof, StorageProof};
 
 pub mod root;
