+++ crates/trie/trie/Cargo.toml
@@ -65,7 +65,7 @@ criterion.workspace = true
 
 [features]
 metrics = ["reth-metrics", "dep:metrics"]
-test-utils = ["triehash"]
+test-utils = ["triehash", "reth-trie-common/test-utils"]
 
 [[bench]]
 name = "prefix_set"
@@ -73,4 +73,9 @@ harness = false
 
 [[bench]]
 name = "hash_post_state"
+harness = false
+
+[[bench]]
+name = "trie_root"
+required-features = ["test-utils"]
 harness = false
