+++ crates/trie/trie/src/test_utils.rs
@@ -1,6 +1,6 @@
 use alloy_rlp::encode_fixed_size;
-use reth_primitives::{proofs::triehash::KeccakHasher, Account, Address, B256, U256};
-use reth_trie_common::TrieAccount;
+use reth_primitives::{Account, Address, B256, U256};
+use reth_trie_common::{triehash::KeccakHasher, TrieAccount};
 
 /// Re-export of [triehash].
 pub use triehash;
